refactor(focus): replace App-level focus handling with recursive widget-based focus system
- Replaced global App focus tracking with per-widget has_focus property
- Implemented recursive Widget_Focus() and Widget_Blur() for hierarchical focus propagation
- Simplified input handling: input now propagates automatically through focused widget branches
- Updated App_HandleInput() to delegate input directly to root widget
- Removed obsolete focus management functions (App_SetFocus, App_ClearFocus, App_HasFocus)
- Minor fixes in Stack (use size_t for loop index, include logging header)

+++ src/common/stack.c
@@ -1,4 +1,5 @@
 #include "stack.h"
+#include "logging.h"
 
 static void increase_capacity(Stack *stack) {
     if (!stack) {
@@ -61,7 +62,7 @@ bool Stack_Has(const Stack *stack, const void *item) {
     if (!stack) {
         return false;
     }
-    for (int i=0; i<stack->size; i++) {
+    for (size_t i=0; i<stack->size; i++) {
         if (stack->items[i] == item) {
             return true;
         }
